+++ main.cpp
@@ -2,13 +2,13 @@
 #include "vector"
 
 #include "B-tree_nodes.hpp"
-#include "B-tree-initial.hpp"
+#include "B-tree-initial.cpp"
 using namespace utlis;
 int main() {
 
 
-    Manager<int,int> manager(1);
-    manager.data_insert(1);
+    middle_node<int,int> * root = utlis::initial_B_tree<int, int>(1);
+    //vector<vector<int>> a;
     int a;
